requirements.txt
Django==1.10.6
djangorestframework==3.5.4

